requirements.txt
Flask
openai
python-dotenv
beautifulsoup4
requests
supabase
